Set up for glob/non-glob

+++ test-glob.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+import contextlib
 import os
 import shutil
 import subprocess
@@ -7,37 +8,42 @@ from pathlib import Path
 from textwrap import dedent
 
 
-def create_testcase(*, n=1024, depth=1):
+@contextlib.contextmanager
+def chdir(path):
+    prev = Path.cwd()
+    try:
+        yield os.chdir(path)
+    finally:
+        os.chdir(prev)
+
+
+def create_testcase(source, *, n=1024, depth=1):
     assert depth == 1, "Not implemented"
 
-    shutil.rmtree(Path('./test'))
+    source.mkdir()
+    with chdir(source):
+        Path('CMakeLists.txt').write_text(dedent(
+            '''
+            cmake_minimum_required(VERSION 3.20)
+            project(glob-test)
+    
+            file(GLOB_RECURSE sources CONFIGURE_DEPENDS "src/**.cpp")
+            add_executable(main ${sources})
+            '''
+        ).lstrip())
 
-    Path('./test').mkdir()
+        Path('src').mkdir()
 
-    Path('./test/CMakeLists.txt').write_text(dedent(
-        '''
-        cmake_minimum_required(VERSION 3.20)
-        project(glob-test)
+        Path('src/main.cpp').write_text(dedent(
+            '''
+            int main () {
+                return 0;
+            }
+            '''
+        ).lstrip())
 
-        file(GLOB_RECURSE sources CONFIGURE_DEPENDS "src/**.cpp")
-        add_executable(main ${sources})
-        '''
-    ).lstrip())
-
-    Path('./test/src').mkdir()
-
-    Path('./test/src/main.cpp').write_text(dedent(
-        '''
-        int main () {
-            return 0;
-        }
-        '''
-    ).lstrip())
-
-    for i in range(1, n):
-        Path(f'./test/src/src{i}.cpp').touch()
-
-    return Path('./test').resolve()
+        for i in range(1, n):
+            Path(f'src/src{i}.cpp').touch()
 
 
 def time_command(cmd):
@@ -64,13 +70,18 @@ def test_overhead(source, build, generator='Ninja'):
 
 
 def main():
-    source = create_testcase(n=1024)
+    source = Path('./test')
+    if source.exists():
+        shutil.rmtree(source)
+
     build = Path('./build')
+    if build.exists():
+        shutil.rmtree(build)
 
     default_generator = ('Visual Studio 16 2019' if os.name == 'nt'
                          else 'Ninja')
 
-    shutil.rmtree(build)
+    create_testcase(source, n=1024)
     overhead = test_overhead(source, build, default_generator)
 
     print(f'Overhead was: {overhead} seconds')
